Further refactoring, added workaround for non-adjustable include paths in Arduino IDE
