+++ src/SolarTracker/Direction.hpp
@@ -1,0 +1,12 @@
+#ifndef DIRECTION_H
+#define DIRECTION_H
+
+namespace SolarTracker
+{
+    enum class Direction : uint8_t
+    {
+        Positive, Negative, Stop
+    };
+}
+
+#endif
